replace getitem in dask/array/chunk.py
