+++ dask/array/utils.py
@@ -10,6 +10,7 @@ from tlz import concat, frequencies
 
 from ..highlevelgraph import HighLevelGraph
 from ..utils import has_keyword, is_arraylike, is_cupy_type
+from .core import Array
 
 try:
     AxisError = np.AxisError
@@ -44,8 +45,6 @@ def meta_from_array(x, ndim=None, dtype=None):
     -------
     array-like with zero elements of the correct dtype
     """
-    from .core import Array
-
     # If using x._meta, x must be a Dask Array, some libraries (e.g. zarr)
     # implement a _meta attribute that are incompatible with Dask Array._meta
     if hasattr(x, "_meta") and isinstance(x, Array):
@@ -240,8 +239,6 @@ def _get_dt_meta_computed(x, check_shape=True, check_graph=True, check_chunks=Tr
     x_meta = None
     x_computed = None
 
-    from .core import Array
-
     if isinstance(x, Array):
         assert x.dtype is not None
         adt = x.dtype
@@ -402,8 +399,6 @@ def _array_like_safe(np_func, da_func, a, like, **kwargs):
     if like is a and hasattr(a, "__array_function__"):
         return a
 
-    from .core import Array
-
     if isinstance(like, Array):
         return da_func(a, **kwargs)
     elif isinstance(a, Array):
@@ -535,13 +530,3 @@ def scipy_linalg_safe(func_name, *args, **kwargs):
 
 def solve_triangular_safe(a, b, lower=False):
     return scipy_linalg_safe("solve_triangular", a, b, lower=lower)
-
-
-def getitem(obj, index):
-    result = obj[index]
-    try:
-        if obj.size >= 2 * result.size:
-            result = result.copy()
-    except AttributeError:
-        pass
-    return result
